+++ ConsumePlugin/GeneratedJson.fs
@@ -4,6 +4,7 @@
 //------------------------------------------------------------------------------
 
 
+
 namespace ConsumePlugin
 
 open System.Text.Json.Serialization