+++ ConsumePlugin/GeneratedRestClient.fs
@@ -6,6 +6,7 @@
 
 
 
+
 namespace PureGym
 
 open System