+++ ConsumePlugin/GeneratedVault.fs
@@ -5,6 +5,7 @@
 
 
 
+
 namespace ConsumePlugin
 
 /// Module containing JSON parsing methods for the JwtVaultAuthResponse type