+++ WoofWare.Myriad.Plugins/Parameters.fs
@@ -2,6 +2,7 @@ namespace WoofWare.Myriad.Plugins
 
 type internal DesiredGenerator =
     | InterfaceMock of isInternal : bool option
+    | CapturingInterfaceMock of isInternal : bool option
     | JsonParse of extensionMethod : bool option
     | JsonSerialize of extensionMethod : bool option
     | HttpClient of extensionMethod : bool option
@@ -11,6 +12,9 @@ type internal DesiredGenerator =
         | "GenerateMock" -> DesiredGenerator.InterfaceMock None
         | "GenerateMock(true)" -> DesiredGenerator.InterfaceMock (Some true)
         | "GenerateMock(false)" -> DesiredGenerator.InterfaceMock (Some false)
+        | "GenerateCapturingMock" -> DesiredGenerator.CapturingInterfaceMock None
+        | "GenerateCapturingMock(true)" -> DesiredGenerator.CapturingInterfaceMock (Some true)
+        | "GenerateCapturingMock(false)" -> DesiredGenerator.CapturingInterfaceMock (Some false)
         | "JsonParse" -> DesiredGenerator.JsonParse None
         | "JsonParse(true)" -> DesiredGenerator.JsonParse (Some true)
         | "JsonParse(false)" -> DesiredGenerator.JsonParse (Some false)